requirements.txt
openai>=1.0.0
pytest>=8.0.0
pytest-bdd>=7.1.2
pytest-playwright>=0.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
jira>=3.5.2
streamlit>=1.30.0
